requirements.txt
google-search-results
langchain-community
gradio